organized layout code into static files
