+++ Braze Demo/View/CompositionalLayout/LayoutBuilder.swift
@@ -1,0 +1,13 @@
+import UIKit
+
+class LayoutBuilder {
+  static func buildAdLayoutSection(layoutEnvironment: NSCollectionLayoutEnvironment) -> NSCollectionLayoutSection {
+    let width = layoutEnvironment.container.effectiveContentSize.width
+    let height: CGFloat = width/394 * 100
+    let size = NSCollectionLayoutSize(widthDimension: NSCollectionLayoutDimension.fractionalWidth(1), heightDimension: NSCollectionLayoutDimension.absolute(height))
+    let item = NSCollectionLayoutItem(layoutSize: size)
+    let group = NSCollectionLayoutGroup.horizontal(layoutSize: size, subitem: item, count: 1)
+    let section = NSCollectionLayoutSection(group: group)
+    return section
+  }
+}
